Move job creation into separate module
