+++ src/SimulationService.jl
@@ -9,12 +9,10 @@ import Oxygen: serveparallel, resetstate, json, setschema, terminate, @post, @ge
 import SwaggerMarkdown: build, @swagger, OpenAPI, validate_spec, openApiToDict, DOCS
 import YAML
 import JSON3 as JSON
-import DataFrames: DataFrame
-import HTTP: Request, Response
 import JobSchedulers: scheduler_start, set_scheduler, scheduler_stop, submit!, job_query, result, generate_id, update_queue!, Job, JobSchedulers
 
-include("./contracts/Interface.jl"); import .Interface: get_operation, use_operation, conversions_for_valid_inputs, Context
-include("./service/Service.jl"); import .Service.ArgIO: prepare_output, prepare_input; import .Service.Queuing: publish_to_rabbitmq
+include("./service/Service.jl")
+import .Service make_deterministic_run, retrieve_job
 include("./Settings.jl"); import .Settings: settings
 
 export start!, stop!
@@ -32,89 +30,13 @@ function health_check()
     return "Simulation-service. TDS_URL=$tds_url, RABBITMQ_ROUTE=$mq_route, ENABLE_REMOTE_DATA_HANDLING=$enable_remote"
 end
 
-"""
-Generate the task to run with the correct context    
-"""
-function contextualize_prog(context)
-    prepare_output(context) ∘ use_operation(context) ∘ prepare_input(context)
-end
-
-"""
-Schedule a sim run given an operation
-"""
-function make_deterministic_run(req::Request, operation::String)
-    # TODO(five): Spawn remote workers and run jobs on them
-    # TODO(five): Handle Python so a probabilistic case can work
-    if isnothing(get_operation(operation))
-        return Response(
-            404,
-            ["Content-Type" => "text/plain; charset=utf-8"],
-            body="Operation not found"
-        )
-    end
-
-    publish_hook = settings["SHOULD_LOG"] ? publish_to_rabbitmq : (args...) -> nothing
-
-    context = Context(
-        generate_id(),
-        publish_hook,  
-        Symbol(operation),
-    )
-    prog = contextualize_prog(context)
-    sim_run = Job(@task(prog(req)))
-    sim_run.id = context.job_id
-    submit!(sim_run)
-    Response(
-        201,
-        ["Content-Type" => "application/json; charset=utf-8"],
-        body=JSON.write("id" => sim_run.id)
-    )
-end
-
-"""
-Get status of sim
-"""
-function retrieve_job(_, id::Int64, element::String)
-    job = job_query(id)
-    if isnothing(job)
-        return Response(
-            404,
-            ["Content-Type" => "text/plain; charset=utf-8"],
-            body="Job does not exist"
-        )
-    end
-    if element == "status"
-        return Dict("status" => job.state)
-    elseif element == "result"
-        if job.state == :done
-            return result(job)
-        else
-            return Response(
-                400,
-                ["Content-Type" => "text/plain; charset=utf-8"],
-                body="Job has not completed"
-            )
-        end
-    else
-        return Response(
-            404,
-            ["Content-Type" => "text/plain; charset=utf-8"],
-            body="Element not found"
-        )
-    end
-end
-
-
-
 
 """
 Specify endpoint to function mappings
 """
 function register!()
     @get "/" health_check
-
-    @get "/runs/{id}/{element}" retrieve_job
-
+    @get "/{element}/{id}" retrieve_job
     @post "/{operation}" make_deterministic_run
 
     info = Dict("title" => "Simulation Service", "version" => "0.6.0")
